Create a separate component for articles
